requirements.txt
apispec
colorama
coloredlogs
falcon
falcon-apispec
falcon-auth
falcon-marshmallow
gitpython
jproperties
pint
psutil
pyyaml
requests
swagger-ui-py
ttldict
verboselogs
waitress
werkzeug
